pyproject.toml
[project]
name = "mnemos"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.18.4",
    "asyncpg>=0.31.0",
    "fastapi[standard]>=0.135.1",
    "httpx>=0.28.1",
    "openai>=2.24.0",
    "pydantic-ai>=1.66.0",
    "pydantic-settings>=2.13.1",
    "sqlalchemy>=2.0.48",
    "streamlit>=1.55.0",
]
